+++ backend/src/main/java/aimo/backend/common/security/dto/JwtDTO.java
@@ -1,4 +1,4 @@
-package aimo.backend.domains.security.dto;
+package aimo.backend.common.security.dto;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
